refactor: split up gatewayconfig_app
- broke gatewayconfig_app into multiple classes, organized by advertisements, descriptors, characteristics, and services
- renamed opaqueStructure and utf8Format to follow python class name conventions.
- fixed some imports, but not all
