+++ gatewayconfig/bluetooth/descriptors/mac_address_descriptor.py
@@ -1,0 +1,19 @@
+import dbus
+from lib.cputemp.service import Descriptor
+import gatewayconfig.constants as constants
+
+class MacAddressDescriptor(Descriptor):
+
+    def __init__(self, characteristic):
+        Descriptor.__init__(
+                self, constants.USER_DESC_DESCRIPTOR_UUID,
+                ["read"],
+                characteristic)
+
+    def ReadValue(self, options):
+        value = []
+        desc = constants.MAC_ADDRESS_VALUE
+
+        for c in desc:
+            value.append(dbus.Byte(c.encode()))
+        return value
