+++ lib/cputemp/advertisement.py
@@ -23,7 +23,7 @@ SOFTWARE.
 import dbus
 import dbus.service
 
-from bletools import BleTools
+from lib.cputemp.bletools import BleTools
 
 BLUEZ_SERVICE_NAME = "org.bluez"
 LE_ADVERTISING_MANAGER_IFACE = "org.bluez.LEAdvertisingManager1"
